requirements.txt
python-decouple==3.6
selenium~=4.3.0
h11~=0.13.0
pip~=21.2.4
attrs~=21.4.0
cryptography~=37.0.4
trio~=0.21.0
outcome~=1.2.0
sniffio~=1.2.0
sortedcontainers~=2.4.0
cffi~=1.15.1
idna~=3.3
certifi~=2022.6.15
six~=1.16.0
urllib3~=1.26.10
pycparser~=2.21
setuptools~=58.1.0
wsproto~=1.1.0
pandas~=1.4.3